split MLIR into a separate tool to have both Core and MLIR printers.
Cannot be in same project because of orphan Pretty instances. Both MLIR
and the existing Core pretty printer would need to implement (Pretty XX)
where XX are the data types we are trying to print. This is not possible
because every data type can have at most one implementation of a
typeclass (typeclasses form a lattice)